Morpheus/requirements.txt
# MORPHEUS System Requirements
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Machine learning and neural networks  
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Physics simulation
pybullet>=3.2.5
trimesh>=3.23.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Configuration and validation
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.65.0

# Visualization (optional)
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.5.0
flake8>=6.0.0

# Optional GPU support
# torch-geometric>=2.4.0  # Uncomment for graph neural networks
# cupy>=12.0.0  # Uncomment for GPU acceleration